The problem and solution
The problem is that once the debugCreateInfo
struct is made the even if the object
that made it, is with us, createInfo pointer
vanishes. Two solutions:
- Either implement state with unique ptr etc
- Or try another design.

I think since we are creating lots of parameters
at compile time. It might be interesting to
manage default values with templates. Something like
template <unsigned int N> struct something gives
me a debugUtility for severe errors etc.