New design and implementation details
We start organizing the api around tasks. The general pattern should
be something hybrid between mediator and state, where the signals
of the tasks are correctly interpreted by the graph object.
Thus a graph holds a two maps:
- node_id -> node
- node_id -> vector whose index is signals and data is node_ids or
  labels depending on the instantiation of the graph template.

A task goes something like this:
- It has a computation that outputs a signal
- It has an identifier and a label
- It has compile time assembled array of const_str of dependencies.
- The dependencies are just labels of antecedent tasks.

A node goes like this:
- It has an identifier and a label
- It has a task.
- It has a compile time assembled array of const_str of dependency
  signals along with the labels of nodes that contain tasks which
  output those signals.

Not sure about the node design it seems like it is invalidating or
repeating the usage of tasks. Essentially tasks create a
spanning tree, and nodes create a finite state machine where the
graph manages transition between states. Not sure how to implement
this structure.