Identified the key problem
The key problem with the design of vulkan is that the tasks
are backward looking, that is, you care about what needs
to be done before it, and states/nodes are forward looking
that is, you care about what comes after it. You concern
yourself with what your returned value is consumed with.
Again, tasks are a tree, states are a graph. The first one
can not contain cycles the second one can.
You can deduce the first from the second one. Thus you don't
need to specify the first one directly.