The code compiles the new style for adding nodes
The new style for adding nodes has been exemplified
in the vkgraphtri.cpp. Basically more or less the same
thing with more information about the type of branching
we have for a particular node. We can serialize then
the resulting tree and work on improving the architecture
more. This application evolves into something that can
be used to improve vulkan applications.