Refactoring continues
The refactoring of app creation continues. The main goal now is to interchange
the insides of functions that are accumulated in vktriapp.hpp with
functions that are delegated to headers.

This has been done with
vkapp.hpp and vkinstance.hpp. Both contain now a maker function which
sets what can be set using templates and delegate what can't be set at
compile time to function arguments. Literal classes are also a good way of
isolating elements that have similar types, see for example
the engine version and app version in vkapp.hpp.