Heavy refactoring
The heavy cache of builds by clang hides the error that were
present in the code base before we had started with the refactoring.

Currently the code compiles, however the scene is problematic.
I am assuming there is a problem in shader creation,
subpass references and subpass attachment and/or attachment
references. The previous segfault location was the
creation of fragment shader. It disappeared but the scene is still
problematic.