New struct setting approach started.
The new struct setting approach that is setting everything
with a designated VkStructSetter instance has started.
This should significantly simplfy the application logic.
We have a single object that sets values to other objects,
instead of multiple wrappers.
We can probably specialize the template so that it would
accept flags as well as other literal values, for setting
bitmask at a compile time.