A new refactoring idea started.
Now the new refactoring idea is very simple. Instead of specializing
templates and dealing with signatures, we simply overload aggressively
3 functions VkFlagSetter which is a constexpr function for loading
required flags. VkOptSetter which is for optional arguments.
VkArgSetter which is for runtime values. This scheme is much simpler,
and should work better for the user and give better error messages.